Cleared size_t/int warning.

+++ src/crypto/PlainText.cpp
@@ -78,7 +78,7 @@ if(toString() == obj.toString())
 return false;
 }
 
-int PlainText::length()
+size_t PlainText::length()
 {
 return rawBytes.length();
 }
